+++ packages/target-tauri/src-tauri/src/html_window/mod.rs
@@ -26,7 +26,7 @@ use crate::{
     },
     state::html_email_instances::InnerHtmlEmailInstanceData,
     temp_file::get_temp_folder_path,
-    DeltaChatAppState, HtmlEmailInstancesState, MenuManger,
+    DeltaChatAppState, HtmlEmailInstancesState, MenuManager, TranslationState,
 };
 
 const HEADER_HEIGHT: f64 = 100.;
@@ -44,7 +44,7 @@ pub(crate) async fn open_html_window(
     app: tauri::AppHandle,
     html_instances_state: State<'_, HtmlEmailInstancesState>,
     dc: State<'_, DeltaChatAppState>,
-    menu_manager: State<'_, MenuManger>,
+    menu_manager: State<'_, MenuManager>,
     window_id: &str,
     account_id: u32, // TODO needs to be used later for fetching webrequests over dc core
     is_contact_request: bool,
@@ -119,6 +119,7 @@ pub(crate) async fn open_html_window(
 
     let app_arc = Arc::new(app);
     let app = app_arc.clone();
+    let app_for_nav = app_arc.clone();
 
     let initial_url = {
         #[cfg(not(any(target_os = "windows", target_os = "android")))]
@@ -139,6 +140,7 @@ pub(crate) async fn open_html_window(
     )
     .disable_javascript()
     .on_navigation(move |url| {
+        let tx: tauri::State<'_, TranslationState> = app_for_nav.state::<TranslationState>();
         if url.to_string() == "about:blank" {
             return true;
         }
@@ -148,7 +150,7 @@ pub(crate) async fn open_html_window(
         // When `false`, it still _might_ get intercepted,
         // but only if the message contains some weird links like
         // `email://other.host/`.
-        // We only really care about navigating to `initial_url`:
+        // We only really care about  navigating to `initial_url`:
         // the HTML message viewer is not supposed to be multipage,
         // so it's OK to handle such weird links as external, below.
         let will_be_intercepted = url.scheme() == initial_url_scheme
@@ -172,12 +174,12 @@ pub(crate) async fn open_html_window(
                 app_arc
                     .dialog()
                     .message(
-                        "Punycode detected: tx('puny_code_warning_question', '$$asciiHostname$$')",
+                        format!("Punycode detected: {}", tx.sync_translate("puny_code_warning_question", /*"puny_code_decode_host(orginal_host_name)"*/)),
                     )
-                    .title("tx('puny_code_warning_header')")
+                    .title(tx.sync_translate("puny_code_warning_header"))
                     .kind(tauri_plugin_dialog::MessageDialogKind::Warning)
                     .buttons(tauri_plugin_dialog::MessageDialogButtons::OkCancelCustom(
-                        // TODO use translation strings, as soon as translations are avaialble in rust backend
+                        // Todo: use tx function here (Problems with Lifetime)
                         "Continue".to_owned(),
                         "cancel".to_owned(),
                     ))