+++ packages/target-tauri/src-tauri/src/settings.rs
@@ -3,6 +3,8 @@ use log::warn;
 use tauri::{AppHandle, Manager};
 use tauri_plugin_store::StoreExt;
 
+use crate::{state::menu_manager::MenuManager, TranslationState};
+
 pub(crate) const CONFIG_FILE: &str = "config.json";
 
 pub(crate) const LOCALE_KEY: &str = "locale";
@@ -19,11 +21,15 @@ pub(crate) const HTML_EMAIL_ALWAYS_ALLOW_REMOTE_CONTENT_DEFAULT: bool = false;
 
 // runtime calls this when desktop settings change
 #[tauri::command]
-pub fn change_desktop_settings_apply_side_effects(app: AppHandle, key: &str) -> Result<(), String> {
+pub async fn change_desktop_settings_apply_side_effects(
+    app: AppHandle,
+    key: &str,
+) -> Result<(), String> {
     match key {
         ZOOM_FACTOR_KEY => apply_zoom_factor(&app),
         // "minimizeToTray" => // TODO
         CONTENT_PROTECTION_KEY => apply_content_protection(&app),
+        LOCALE_KEY => apply_language_change(&app).await,
         _ => Ok(()),
     }
     .map_err(|err| format!("{err:#}"))
@@ -106,6 +112,14 @@ pub(crate) fn get_content_protection(app: &AppHandle) -> bool {
             CONTENT_PROTECTION_DEFAULT
         }
     }
+}
+
+pub(crate) async fn apply_language_change(app: &AppHandle) -> anyhow::Result<()> {
+    app.state::<TranslationState>()
+        .reload_from_config(&app)
+        .await?;
+    app.state::<MenuManager>().update_all(&app);
+    Ok(())
 }
 
 pub(crate) fn get_setting_bool_or(